backend/requirements.txt
setuptools>=69.0
django==5.0
djangorestframework==3.15
django-cors-headers==4.3
djangorestframework-simplejwt==5.4.0
channels[daphne]==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
django-environ==0.11
celery==5.3
redis==5.0
django-celery-beat==2.6
django-storages==1.14
boto3==1.34
openfoodfacts==1.1
Pillow==10.2
django-filter==23.5
drf-spectacular==0.27
pywebpush==2.0
gunicorn==21.2
whitenoise==6.6
